% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sysreq.R
\name{format_sysreq}
\alias{format_sysreq}
\title{Cross reference the \code{pkg} against
a custom Google sheet to get formatted
system requirements}
\usage{
format_sysreq(pkg, sysreqs_db)
}
\arguments{
\item{pkg}{is the package name to be
cross referenced with the Neuroconductor
list of formatted system requirements, it
can also be a repo slug}

\item{sysreqs_db}{is the data.frame containing
the Google sheet system requirements list}
}
\value{
a list of formatted system requirements
 if the package is in the Neuroconductor system
 requirements list
}
\description{
Cross reference the \code{pkg} against
a custom Google sheet to get formatted
system requirements
}
