% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sysreq.R
\name{get_all_sysreqs}
\alias{get_all_sysreqs}
\title{Get a list of all system requirements for a package
using the recursive dependencies tree}
\usage{
get_all_sysreqs(pkg)
}
\arguments{
\item{pkg}{repo slug}
}
\value{
a list with all system dependencies for \code{pkg}
}
\description{
Get a list of all system requirements for a package
using the recursive dependencies tree
}
\examples{
\dontrun{get_all_sysreqs('neuroconductor/ANTsR')}

}
