% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sysreq.R
\name{get_dependency_tree}
\alias{get_dependency_tree}
\title{Gets dependency tree for a package}
\usage{
get_dependency_tree(pkg, cran = TRUE, bioc = TRUE, URLs = NULL)
}
\arguments{
\item{pkg}{character vector of packages}

\item{cran}{search CRAN}

\item{bioc}{search Bioconductor}

\item{URLs}{URLs for third party package DBs}
}
\value{
a list of unique package names in the dependency
   tree
}
\description{
\code{get_dependency_tree} will build and return
a dependency tree. This is done recursively
for the import/depends fields and non-recursively
for the suggests field.
}
\examples{
\dontrun{get_dependency_tree('ggplot2',bioc=FALSE)}
\dontrun{get_dependency_tree('fslr',URLs='http://neuroconductor.org:8080')}

}
