% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sysreq.R
\name{get_gh_repo_dependencies}
\alias{get_gh_repo_dependencies}
\title{Gets the import/depends/suggest list from
a GitHub repo}
\usage{
get_gh_repo_dependencies(slug)
}
\arguments{
\item{slug}{is the GitHub repository name
in the form of username/repo}
}
\value{
a list of unique package names that
  \code{slug} repo depends on, from the
  DESCRIPTION file. This is a combination
  of the Imports, Depends and Suggests
  tags.
}
\description{
\code{get_gh_repo_dependencies} will parse
 a remote GitHub repo's DESCRIPTION file
 and it will return a list of unique
 dependencies (from Imports/Dependes/Suggests)
}
\examples{
\dontrun{get_gh_repo_dependencies('neuroconductor/fslr')}

}
