% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sysreq.R
\name{get_single_pkg_sysreqs}
\alias{get_single_pkg_sysreqs}
\title{Get repo SystemRequirements}
\usage{
get_single_pkg_sysreqs(pkg, sysreqs_db)
}
\arguments{
\item{pkg}{is the package name to be
cross referenced with the Neuroconductor
list of formatted system requirements, it
can also be a repo slug}

\item{sysreqs_db}{is the data.frame containing
the Google sheet system requirements list}
}
\value{
a list of formatted system requirements
 if the package is in the Neuroconductor system
 requirements list or the original system requirements
 for the DESCRIPTION file (NA if none)
}
\description{
\code{get_single_pkg_sysreqs} returns the
 system requirements for a package by
 cross referencing the package name with
 the Neuroconductor maintained list of
 formatted system requirements
}
\examples{
\dontrun{get_gh_repo_sysreqs('neuroconductor/ANTsR')}
\dontrun{get_gh_repo_sysreqs('fslr')}

}
