% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sysreq.R
\name{read_bioc}
\alias{read_bioc}
\title{Read the Bioconductor package data
to be used with miniCRAN:pkgDep}
\usage{
read_bioc()
}
\description{
Read the Bioconductor package data
to be used with miniCRAN:pkgDep
}
